requirements.txt
adal==0.5.0
asn1crypto==0.24.0
astroid==1.6.1
azure-common==1.1.8
azure-mgmt-network==1.7.1
azure-mgmt-nspkg==2.0.0
azure-nspkg==2.0.0
backports.functools-lru-cache==1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
configparser==3.5.0
cryptography==2.1.4
enum34==1.1.6
futures==3.2.0
idna==2.6
ipaddress==1.0.19
isodate==0.6.0
isort==4.3.1
keyring==11.0.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
msrest==0.4.26
msrestazure==0.4.21
oauthlib==2.0.6
pycparser==2.18
PyJWT==1.5.3
pylint==1.8.2
python-dateutil==2.6.1
requests==2.18.4
requests-oauthlib==0.8.0
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.22
wrapt==1.10.11
